modified:   delivery/src/main/java/mall/domain/Delivery.java 	modified:   delivery/src/main/java/mall/domain/DeliveryCanceled.java 	modified:   delivery/src/main/java/mall/domain/DeliveryCompleted.java 	modified:   delivery/src/main/java/mall/domain/DeliveryRepository.java 	modified:   delivery/src/main/java/mall/domain/DeliveryReturned.java 	renamed:    delivery/src/main/java/mall/domain/DeleveryStarted.java -> delivery/src/main/java/mall/domain/DeliveryStarted.java 	modified:   init.sh 	new file:   kubectl 	renamed:    product/src/main/java/mall/domain/DeliveryReturned.java -> product/src/main/java/mall/domain/DeliveryCanceled.java 	renamed:    product/src/main/java/mall/domain/DeliveryCompleted.java -> product/src/main/java/mall/domain/DeliveryStarted.java 	modified:   product/src/main/java/mall/domain/Inventory.java 	modified:   product/src/main/java/mall/infra/PolicyHandler.java

+++ delivery/src/main/java/mall/domain/DeliveryCompleted.java
@@ -13,7 +13,7 @@ public class DeliveryCompleted extends AbstractEvent {
     private Long id;
     private Long orderId;
     private Long productId;
-    private Long qty;
+    private Integer qty;
     private String productName;
     private String status;
